Separates Course class
Made __init__  into holding the course class with only the global vars and made an Course_create class, which contains the create function
