--- .setup/bin/sample_courses/models/course/global_var.py
@@ -1,19 +1,0 @@
-
-from sqlalchemy import Table
-
-
-
-class Table_var:
-    # This is not really needed but should stay for readability
-    gradeable_table: Table
-    electronic_table: Table
-    peer_assign: Table
-    reg_table: Table
-    component_table: Table
-    mark_table: Table
-    gradeable_data: Table
-    gradeable_component_data: Table
-    gradeable_component_mark_data: Table
-    gradeable_data_overall_comment: Table
-    electronic_gradeable_data: Table
-    electronic_gradeable_version: Table
